# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthorInfo:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
...
  dto.ChangeRoleRequest:
    properties:
      role:
        enum:
        - user
        - admin
        type: string
    required:
    - role
    type: object
  dto.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/dto.AuthorInfo'
      content:
        type: string
      content_html:
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      post_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
    type: object
  dto.CreateCommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
...
  dto.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      meta:
        $ref: '#/definitions/dto.Meta'
      success:
        type: boolean
    type: object
  dto.SignupRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.UpdateCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 게시판 API 서버입니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go Board API
  version: "1.0"
paths:
  /admin/posts/{id}:
    delete:
      description: 관리자가 게시글을 강제 삭제합니다
      parameters:
      - description: 게시글 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 게시글 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 게시글 강제 삭제
      tags:
      - admin
  /admin/stats:
    get:
      description: 전체 사이트 통계를 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 관리자 통계 조회
      tags:
      - admin
  /admin/users:
    get:
      description: 관리자용 사용자 목록을 조회합니다
      parameters:
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지 크기
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 사용자 목록 조회
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: 관리자가 사용자를 삭제합니다
      parameters:
      - description: 사용자 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 사용자 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 사용자 삭제
      tags:
      - admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: 사용자의 역할을 변경합니다
      parameters:
      - description: 사용자 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 역할 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 사용자 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 사용자 역할 변경
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: 이메일과 비밀번호로 로그인합니다
      parameters:
      - description: 로그인 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 로그인
      tags:
      - auth
  /auth/logout:
    post:
      description: 현재 세션을 종료합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 로그아웃
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh Token으로 Access Token을 갱신합니다
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 토큰 만료
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 토큰 갱신
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: 새 계정을 생성합니다
      parameters:
      - description: 회원가입 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 유효성 검사 실패
          schema:
            $ref: '#/definitions/dto.Response'
        "409":
          description: 이메일 중복
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 회원가입
      tags:
      - auth
  /comments/{id}:
    delete:
      description: 댓글을 삭제합니다
      parameters:
      - description: 댓글 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 댓글 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 댓글 삭제
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: 댓글을 수정합니다
      parameters:
      - description: 댓글 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 수정 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 댓글 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 댓글 수정
      tags:
      - comments
...
  /posts/{postId}/comments:
    get:
      description: 게시글의 댓글 목록을 조회합니다
      parameters:
      - description: 게시글 ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 게시글 없음
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 댓글 목록 조회
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: 게시글에 댓글을 작성합니다
      parameters:
      - description: 게시글 ID
        in: path
        name: postId
        required: true
        type: integer
      - description: 댓글 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: 유효성 검사 실패
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 게시글 없음
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 댓글 작성
      tags:
      - comments
...
securityDefinitions:
  Bearer:
    description: 'JWT 인증 토큰. 형식: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...